Need double for template
Former-commit-id: 47ad3b2b643dc72601df8fdc52230f9f7fa31882

+++ analysis/SamplerAnalysis.hh
@@ -21,7 +21,7 @@ public:
 		  bool debugIn = false);
 #else 
   SamplerAnalysis(BDSOutputROOTEventSampler<double>* samplerIn,
-		  BDSOutputROOTEventSampler<float>* samplerFirst,
+		  BDSOutputROOTEventSampler<double>* samplerFirst,
 		  bool debugIn = false);
 #endif
   /// Initialisation of arrays for optical function calculations
